chore(execute): move step_edit into megaplan/execute/ and remove execution_timeout shim
Phase 2 Step 9 of the file-organization plan: consolidate the remaining
execute-group helpers under megaplan/execute/.

- Delete megaplan/execution_timeout.py (was already a thin re-export of
  megaplan.execute.timeout). Rewrite its single caller tests/test_doc_mode.py
  to import directly from megaplan.execute.timeout.
- git mv megaplan/step_edit.py -> megaplan/execute/step_edit.py and
  atomically rewrite all three call sites (megaplan/__init__.py,
  megaplan/cli.py, megaplan/handlers/shared.py) to use the new path.
  handle_step and next_plan_artifact_name both remain importable from
  megaplan.execute.step_edit; megaplan.handle_step still resolves.

No compatibility shims left behind (U2). Behavior preserved.
