+++ megaplan/__init__.py
@@ -32,7 +32,7 @@ from megaplan.handlers import (
     handle_execute,
     handle_review,
 )
-from megaplan.step_edit import handle_step
+from megaplan.execute.step_edit import handle_step
 from megaplan.flags import (
     normalize_flag_record,
     update_flags_after_critique,
