--- megaplan/execution_timeout.py
@@ -1,17 +1,0 @@
-"""Compatibility wrapper for historical ``megaplan.execution_timeout`` imports."""
-
-from megaplan.execute.timeout import (
-    _merge_timeout_checkpoint,
-    _recover_execute_timeout,
-    _reset_timeout_invalid_tasks,
-    _resolve_execute_approval_mode,
-    _timeout_checkpoint_path,
-)
-
-__all__ = [
-    "_merge_timeout_checkpoint",
-    "_recover_execute_timeout",
-    "_reset_timeout_invalid_tasks",
-    "_resolve_execute_approval_mode",
-    "_timeout_checkpoint_path",
-]
