+++ tests/test_doc_mode.py
@@ -14,7 +14,7 @@ import megaplan._core.io as io_module
 import megaplan.cli
 from megaplan.doc_assembly import assemble_doc, extract_sections
 from megaplan.evaluation import validate_execution_evidence
-from megaplan.execution_timeout import _merge_timeout_checkpoint, _reset_timeout_invalid_tasks
+from megaplan.execute.timeout import _merge_timeout_checkpoint, _reset_timeout_invalid_tasks
 from megaplan.schemas import SCHEMAS, get_execution_schema_key, strict_schema
 from megaplan.types import CliError
